refactor(scroll-sync): extract storage logic into services/storage-service.js
